implemented and tested non-returning random walk strategy

+++ x_view_core/src/x_view.cc
@@ -204,7 +204,7 @@ bool XView::localize(const Graph& query_graph, Eigen::Vector3d* position) {
 
   struct WeightedPosition {
     Eigen::Vector3d position;
-    float weight;
+    double weight;
   };
 
   std::vector<WeightedPosition> matched_positions;
@@ -223,7 +223,7 @@ bool XView::localize(const Graph& query_graph, Eigen::Vector3d* position) {
   }
 
   Eigen::Vector3d estimated_position = Eigen::Vector3d::Zero();
-  float total_weight = 0.f;
+  double total_weight = 0.0;
 
   for(const WeightedPosition& w_p : matched_positions) {
     estimated_position += w_p.position * w_p.weight;
